% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{determineEstimableGenes}
\alias{determineEstimableGenes}
\title{Determine estimable genes using a local polynomial regression
on model components}
\usage{
determineEstimableGenes(outcome, fit, quantile_threshold = 0.75)
}
\arguments{
\item{outcome}{A cell by gene estimated expression matrix}

\item{fit}{Model fit glmnet object}

\item{quantile_threshold}{Percentile at which to threshold estimate standard deviation, default 75\%}
}
\value{
Returns a character vector of the estimable genes
}
\description{
Determine estimable genes using a local polynomial regression
on model components
}
