% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvc}
\alias{mvc}
\title{Single cell RNA-seq and imaging data}
\format{A list
\describe{
  \item{seqFISHplus}{list containing the scRNA-seq expression matrix and additional information}
  \item{scRNAseq}{list containing the scRNA-seq expression matrix}
  ...
}}
\source{
\url{https://www.nature.com/articles/s41586-019-1049-y}, \url{https://www.nature.com/articles/nn.4216}
}
\usage{
mvc
}
\description{
A dataset containing seqFISH+ and scRNA-seq expression matrices
}
\keyword{datasets}
