% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{predictExpression}
\alias{predictExpression}
\title{Train a multi-response elastic-net regression for prediction using
cross-validation to select lambda}
\usage{
predictExpression(training_x, training_y, prediction_x,
  lambda_mse = TRUE, family = "mgaussian", ...)
}
\arguments{
\item{training_x}{A cell by gene expression matrix of training predictor data}

\item{training_y}{A cell by gene expression matrix of training outcome data}

\item{prediction_x}{A cell by gene expression matrix of testing predictor data}

\item{lambda_mse}{Whether to select the lambda that minimizes the cross-validation, default TRUE}

\item{family}{family parameter passed on to the glmnet function, default 'mgaussian'}

\item{...}{additional parameters to pass on to the glmnet function}
}
\value{
Returns a list of the predicted outcome and model fit glmnet object
}
\description{
Train a multi-response elastic-net regression for prediction using
cross-validation to select lambda
}
