% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{quantileNormalize}
\alias{quantileNormalize}
\title{Quantile-normalization of expression data}
\usage{
quantileNormalize(seqFISH_expression, scRNAseq_expression,
  normalize_scRNAseq = TRUE)
}
\arguments{
\item{seqFISH_expression}{A cell by gene expression matrix of testing data}

\item{scRNAseq_expression}{A cell by gene expression matrix of training data data}

\item{normalize_scRNAseq}{Whether to first quantile-normalize the scRNA-seq data, default TRUE}
}
\value{
Returns a list of the normalized matrices
}
\description{
Quantile-normalization of expression data
}
